requirements.txt
accelerate==0.26.1
aiohttp==3.9.1
aiolimiter==1.1.0
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
appnope==0.1.3
asttokens==2.4.1
attrs==23.2.0
backoff==2.2.1
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
cohere==4.44
comm==0.2.1
contourpy==1.2.0
cycler==0.12.1
Cython==0.29.37
debugpy==1.8.0
decorator==5.1.1
distro==1.9.0
einops==0.7.0
executing==2.0.1
fastavro==1.9.3
filelock==3.13.1
Flask==3.0.0
Flask-Cors==4.0.0
fonttools==4.47.2
frozenlist==1.4.1
fsspec==2023.12.2
h11==0.14.0
hdbscan==0.8.33
httpcore==1.0.2
httpx==0.26.0
huggingface-hub==0.20.2
idna==3.6
importlib-metadata==6.11.0
ipykernel==6.29.0
ipython==8.20.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
jupyter_client==8.6.0
jupyter_core==5.7.1
kiwisolver==1.4.5
llvmlite==0.41.1
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
mpmath==1.3.0
multidict==6.0.4
nest-asyncio==1.5.9
networkx==3.2.1
numba==0.58.1
numpy==1.26.3
openai==1.8.0
opt-einsum==3.3.0
packaging==23.2
pandas==2.1.4
parso==0.8.3
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.1.0
prompt-toolkit==3.0.43
psutil==5.9.7
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==14.0.2
pydantic==2.5.3
pydantic_core==2.14.6
Pygments==2.17.2
pynndescent==0.5.11
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.2
regex==2023.12.25
requests==2.31.0
safetensors==0.4.1
scikit-learn==1.3.2
scipy==1.11.4
six==1.16.0
sniffio==1.3.0
sseclient-py==1.8.0
stack-data==0.6.3
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
threadpoolctl==3.2.0
tiktoken==0.5.2
together==0.2.10
tokenizers==0.15.0
torch==2.1.2
tornado==6.4
tqdm==4.66.1
traitlets==5.14.1
transformers==4.36.2
typer==0.9.0
typing_extensions==4.9.0
tzdata==2023.4
umap-learn==0.5.5
urllib3==2.1.0
voyageai==0.1.6
wcwidth==0.2.13
Werkzeug==3.0.1
yarl==1.9.4
zipp==3.17.0
